refactor: [navigation] extract navigator into its own component
