+++ src/navigation/index.ts
@@ -1,0 +1,3 @@
+import Navigator from './Navigator';
+
+export default Navigator;
